refactor: useSearchLogic 훅 개선
