+++ apps/client/src/pages/search/hooks/use-search-logic.ts
@@ -1,10 +1,13 @@
-import { useState } from 'react';
+import { useCallback, useState } from 'react';
 import { useNavigate, useSearchParams } from 'react-router-dom';
 
 export const useSearchLogic = () => {
   const [barFocus, setBarFocus] = useState(false);
+  const [searchParams, setSearchParams] = useSearchParams();
   const navigate = useNavigate();
-  const [, setSearchParams] = useSearchParams();
+  const paramsKeyword = searchParams.get('q') || '';
+  const selectedArtistId = searchParams.get('aid');
+  const selectedPerformanceId = searchParams.get('pid');
 
   const handleNavigateWithKeyword = (keyword: string) => {
     if (!keyword.trim()) return;
@@ -13,15 +16,37 @@ export const useSearchLogic = () => {
     setBarFocus(false);
   };
 
+  const handleSelectKeyword = useCallback(
+    (keyword: string, id: string | number, type: 'artist' | 'performance') => {
+      const newParams = new URLSearchParams(searchParams);
+      newParams.set('q', keyword);
+
+      if (type === 'artist') {
+        newParams.set('aid', String(id));
+        newParams.delete('pid');
+      } else if (type === 'performance') {
+        newParams.set('pid', String(id));
+        newParams.delete('aid');
+      }
+      setSearchParams(newParams);
+    },
+    [searchParams, setSearchParams],
+  );
+
   const handleOnFocus = () => setBarFocus(true);
   const handleOnBlur = () => setBarFocus(false);
   const handleClear = () => setSearchParams(new URLSearchParams());
 
   return {
+    searchParams,
+    paramsKeyword,
+    selectedArtistId,
+    selectedPerformanceId,
     barFocus,
+    handleNavigateWithKeyword,
+    handleSelectKeyword,
     handleOnFocus,
     handleOnBlur,
-    handleNavigateWithKeyword,
     handleClear,
   };
 };
